refactor: break up large file
